fix(component): clean up code, remove logs, and extract mocks
