+++ apps/docs/src/components/UI/api-keys/components/Notification.tsx
@@ -7,8 +7,8 @@ import {
     AlertCircle,
 } from 'lucide-react';
 import { motion } from 'framer-motion';
-import { cn } from '../../../../utils/cn';
-import { Typography } from "../../typography";
+import { cn } from '@site/src/utils/cn';
+import { Typography } from "@site/src/components/UI/typography";
 import { useEffect } from "react";
 import { NotificationVariants } from "../utils";
