+++ apps/docs/src/components/UI/api-keys/components/StatsOverview.tsx
@@ -7,8 +7,8 @@ import {
     CheckCircle,
 } from 'lucide-react';
 import { motion } from 'framer-motion';
-import { cn } from '../../../../utils/cn';
-import { Typography } from "../../typography";
+import { cn } from '@site/src/utils/cn';
+import { Typography } from "@site/src/components/UI/typography";
 import { NewBadge } from "./NewBadge";
 
 export const StatsOverview = ({ stats, isLoading, badgeVariant = "tinypop" }: StatsOverviewProps) => {
