+++ apps/storybook/src/templates/pages/account-management/api-keys/components/ApiKeyCard.tsx
@@ -28,8 +28,7 @@ export const ApiKeyCard = ({
     onCopy,
     onRevoke,
     // onRegenerate,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    showActions = true,
+    // showActions = true,
     variant = 'default',
     badgeVariant = "tinypop",
     buttonVariant = "ghost",
@@ -126,7 +125,7 @@ export const ApiKeyCard = ({
 
                     <div className="relative" onClick={(e) => e.stopPropagation()}>
                         <Button
-                            variant={buttonVariant as any}
+                            variant={buttonVariant}
                             size="icon"
                             onClick={() => setIsMenuOpen(!isMenuOpen)}
                             className="cursor-pointer"
@@ -157,15 +156,6 @@ export const ApiKeyCard = ({
                                     </button>
                                     {apiKey.status === 'active' && (
                                         <>
-                                            {/* {onRegenerate && (
-                                                <button
-                                                    onClick={handleRegenerate}
-                                                    className="flex items-center gap-2 w-full px-3 py-2 text-sm hover:bg-secondary transition-colors cursor-pointer"
-                                                >
-                                                    <RefreshCw className="w-4 h-4" />
-                                                    Regenerate
-                                                </button>
-                                            )} */}
                                             <button
                                                 onClick={handleRevoke}
                                                 className="flex items-center gap-2 w-full px-3 py-2 text-sm text-warning hover:bg-warning/10 transition-colors cursor-pointer"
@@ -198,7 +188,7 @@ export const ApiKeyCard = ({
                             {apiKey.keyPrefix}••••••••{apiKey.keySuffix}
                         </code>
                         <Button
-                            variant={buttonVariant as any}
+                            variant={buttonVariant}
                             size="sm"
                             onClick={(e) => {
                                 e.stopPropagation();
