+++ apps/storybook/src/templates/pages/account-management/api-keys/index.tsx
@@ -19,7 +19,7 @@ import type {
     ApiKey, ApiKeyScope, ApiKeysPageProps, FilterOptions, NotificationType, StatsData
 } from './types';
 import { animationVariants, CardVariants, PageVariants, TableVariants } from './utils';
-import { generateMockApiKeys } from './constants';
+import { generateMockApiKeys } from './mock';
 import { Notification as NotificationComponent } from './components/Notification';
 import { StatsOverview } from './components/StatsOverview';
 import { SearchFilter } from './components/SearchFilter';
@@ -66,14 +66,11 @@ export const ApiKeysPage: React.FC<ApiKeysPageProps> = ({
     showExport = true,
     showStats = true,
     // allowRegeneration = true,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    requireConfirmation = true,
+    // requireConfirmation = true,
     showNotifications = true,
     notificationDuration = 3000,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    requirePasswordToReveal = false,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    autoHideRevealedKey = true,
+    // requirePasswordToReveal = false,
+    // autoHideRevealedKey = true,
     autoHideDelay = 30,
     darkMode = false
 }) => {
@@ -370,7 +367,7 @@ export const ApiKeysPage: React.FC<ApiKeysPageProps> = ({
                             )}
                             <Button
                                 onClick={() => setIsGenerateModalOpen(true)}
-                                variant={buttonVariant as any}
+                                variant={buttonVariant}
                                 className="cursor-pointer"
                                 animationVariant={buttonAnimationVariant}
                             >
