+++ packages/registry/templates/pages/api-keys/modals/GenerateKeyModal.tsx
@@ -7,14 +7,16 @@ import {
     Loader2,
     CheckCircle,
 } from 'lucide-react';
-import { Button } from '../../../../components/button';
-import { AnimatedInput } from '../../../../components/input';
-import { Typography } from '../../../../components/typography';
+import { AnimatedInput } from '@ignix-ui/input';
+import { cn } from '../../../../utils/cn';
+import { Button } from '@ignix-ui/button';
+import { Typography } from '@ignix-ui/typography';
+
 import type { ApiKey, ApiKeyScope, GenerateKeyModalProps } from '../types';
 import { useState } from 'react';
 import { StatusBadge } from '../components/StatusBadge';
 import { SCOPES } from '../constants';
-import { cn } from '../../../../utils/cn';
+// import { cn } from '@site/src/utils/cn';
 import { ScopeBadge } from '../components/ScopeBadge';
 
 export const GenerateKeyModal = ({
@@ -240,7 +242,7 @@ export const GenerateKeyModal = ({
                                         <Button
                                             onClick={handleGenerate}
                                             disabled={isLoading || !formData.name.trim()}
-                                            variant={buttonVariant as any}
+                                            variant={buttonVariant}
                                             animationVariant={buttonAnimationVariant}
                                             className="cursor-pointer"
                                         >
